+++ RELEASENOTES.md
@@ -32,6 +32,8 @@
 *   HLS extension:
     *   Support X-PLAYOUT-LIMIT with HLS interstitials
         ([#2484](https://github.com/androidx/media/issues/24884)).
+    *   Support the `QUERYPARAM` attribute with `#EXT-X-DEFINE`
+        ([#1854](https://github.com/androidx/media/issues/1854)).
 *   DASH extension:
 *   Smooth Streaming extension:
 *   RTSP extension: